Update input variable type in PID function to long for improved precision in calculations.

+++ ROSArduinoBridge/diff_controller.h
@@ -68,7 +68,7 @@ void resetPID(){
 void doPID(SetPointInfo * p) {
   long Perror;
   long output;
-  int input;
+  long input;
 
   //Perror = p->TargetTicksPerFrame - (p->Encoder - p->PrevEnc);
   input = p->Encoder - p->PrevEnc;
